Inject script links into head section
