+++ lib/easycharts.js
@@ -1,0 +1,220 @@
+'use strict';
+
+var injector = require('connect-injector');
+
+var snippet = '<script src="https://unpkg.com/echarts@3.3.2/dist/echarts.min.js"></script>';
+
+var patterns = {
+  rowDelimiter: /\n/,
+  columnDelimiter: /\s*\|\s*/,
+  horizontalHeadings: /^(([^\|]+\|)*([^\|]+)\n)((\s*\d+\s*\|)*(\s*\d+\s*))$/,
+  verticalHeadings: /^(([^\|]+\|)([^\|]+)\n)(([^\|]+\|)(\s*\d+\s*)\n)*(([^\|]+\|)(\s*\d+\s*))$/,
+  twoDimensional: /^(([^\|]+\|)+([^\|]+)\n)(([^\|]+\|)(\s*\d+\s*\|)+(\s*\d+\s*)\n)*(([^\|]+\|)(\s*\d+\s*\|)+(\s*\d+\s*))$/
+};
+
+var parseChartTagContent = function(content) {
+  var lines = content.split(patterns.rowDelimiter).filter(function(line) { return line; });
+  var headers = [];
+  var bodys = [];
+
+  if (patterns.horizontalHeadings.test(content)) {
+    headers = lines[0].split(patterns.columnDelimiter);
+    var data = lines[1].split(patterns.columnDelimiter);
+    bodys.push({ name: 'Value', data: data });
+  } else if (patterns.verticalHeadings.test(content)) {
+    var data = [];
+    for (var i = 1; i < lines.length; i++) {
+      var pair = lines[i].split(patterns.columnDelimiter);
+      headers.push(pair[0]);
+      data.push(pair[1]);
+    }
+
+    var name = lines[0].split(patterns.columnDelimiter)[1];
+    bodys.push({ name: name, data: data });
+  } else if (patterns.twoDimensional.test(content)) {
+    headers = lines[0].split(patterns.columnDelimiter).slice(1);
+    for (var i = 1; i < lines.length; i++) {
+      var line = lines[i].split(patterns.columnDelimiter);
+      var name = line[0];
+      var data = line.slice(1);
+      bodys.push({ name: name, data: data });
+    }
+  }
+
+  var table = {
+    headers: headers,
+    bodys: bodys
+  };
+
+  return table;
+};
+
+var shouldCombined = function(name) {
+ return /^.+\((.+)\)$/.test(name);
+};
+
+var barchartOption = function(title, table, asLinechart) {
+  var option = {
+    title: {
+      text: title
+    },
+    animation: true,
+    tooltip: {
+      trigger: 'axis'
+    },
+    xAxis: {
+      data: table.headers
+    },
+    yAxis: {},
+    series: []
+  };
+
+  var type = asLinechart? 'line' : 'bar';
+  var legend = [];
+  var bodys = table.bodys;
+  for (var i = 0; i < bodys.length; i++) {
+    var row = bodys[i];
+    var name = row['name'];
+    var data = row['data'];
+
+    var item = { name: name, type: type, data: data };
+    if (shouldCombined(name)) {
+      item['stack'] = RegExp.$1;
+    }
+
+    if (asLinechart) {
+      item['areaStyle'] = { normal: {} };
+    }
+
+    legend.push(name);
+    option['series'].push(item);
+  }
+
+  if (legend.length > 1) {
+    option['legend'] = { data: [] };
+    option['legend']['data'] = legend;
+  }
+
+  if (asLinechart) {
+    option['grid'] = {
+      left: '3%',
+      right: '4%',
+      bottom: '3%',
+      containLabel: true
+    };
+    option['xAxis']['boundaryGap'] = false;
+  }
+
+  return option;
+};
+
+var linechartOption = function(title, table) {
+  return barchartOption(title, table, true);
+};
+
+var piechartOption = function(title, table) {
+  var headers = table.headers;
+
+  var option = {
+    animation: true,
+    tooltip: {
+      trigger: 'item',
+      formatter: "{a} <br/>{b}: {c} ({d}%)"
+    },
+    legend: {
+      data: headers
+    },
+    series:[]
+  };
+
+  var row = table.bodys[0];
+  var name = row['name'];
+  var data = row['data'];
+
+  var pairs = [];
+  for (var i = 0; i < data.length; i++) {
+    pairs.push({ value: data[i], name: headers[i] });
+  }
+
+  var item = {
+    name: name,
+    type: 'pie',
+    radius : '55%',
+
+    data: pairs,
+    itemStyle: {
+      emphasis: {
+        shadowBlur: 10,
+        shadowOffsetX: 0,
+        shadowColor: 'rgba(0, 0, 0, 0.5)'
+      }
+    }
+  };
+
+  option['series'].push(item);
+
+  return option;
+};
+
+var chartOption = function(type, title, table) {
+  var option = {};
+  if (type === 'bar') {
+    option = barchartOption(title, table);
+  } else if (type === 'line') {
+    option = linechartOption(title, table);
+  } else if (type === 'pie') {
+    option = piechartOption(title, table);
+  }
+
+  return option;
+};
+
+var chartTag = function(type, args, content) {
+  var html = content;
+  var title = args[0] || '';
+  var table = parseChartTagContent(content);
+  var option = chartOption(type, title, table);
+
+  if (option['series'] && option['series'].length > 0) {
+    var tagId = Math.random().toString(36).slice(2);
+    var html = `
+      <div id="${tagId}" class="hexo-tag-easy-charts" style="width: 100%; height: 480px;"></div>
+
+      <script type="text/javascript">
+        var chart = echarts.init(document.getElementById('${tagId}'));
+        chart.setOption(${JSON.stringify(option)});
+      </script>
+    `
+  }
+
+  return html;
+};
+
+module.exports = {
+  linksInjector: function(app) {
+    app.use(injector(function(req, res){
+      var contentType = res.getHeader('content-type');
+      return contentType && (contentType.toLowerCase().indexOf('text/html') >= 0);
+    }, function(content, req, res, callback) {
+      var str = content.toString();
+      var pos = str.lastIndexOf('</head>');
+      if (!~pos) {
+        return callback(null, content);
+      }
+      var result = str.substring(0, pos) + snippet + str.substring(pos);
+      callback(null, result);
+    }));
+  },
+
+  barchartTag: function(args, content) {
+    return chartTag('bar', args, content);
+  },
+
+  linechartTag: function(args, content) {
+    return chartTag('line', args, content);
+  },
+
+  piechartTag: function(args, content) {
+    return chartTag('pie', args, content);
+  }
+};
